Split message suggestion sources
