+++ src/features/board/suggestions/suggestion-source.ts
@@ -1,0 +1,13 @@
+import type { Status } from "@shayc/react-built-in-ai";
+
+export interface SuggestionSource {
+  engineStatus: Status;
+  candidate: string | undefined;
+  isRequestPending: boolean;
+  hasRequestFailed: boolean;
+  prepare: () => void;
+}
+
+export function isRequestFailure(error: Error | undefined): boolean {
+  return error !== undefined && error.name !== "AbortError";
+}
